+++ temba/api/tasks.py
@@ -2,13 +2,13 @@ from django.conf import settings
 from django.utils import timezone
 
 from temba.utils import chunk_list
-from temba.utils.celery import nonoverlapping_task
+from temba.utils.crons import cron_task
 
 from .models import WebHookEvent
 
 
-@nonoverlapping_task(name="trim_webhook_event_task")
-def trim_webhook_event_task():
+@cron_task()
+def trim_webhook_events():
     """
     Trims old webhook events
     """