+++ temba/channels/tasks.py
@@ -12,21 +12,21 @@ from temba import mailroom
 from temba.orgs.models import Org
 from temba.utils import chunk_list
 from temba.utils.analytics import track
-from temba.utils.celery import nonoverlapping_task
+from temba.utils.crons import cron_task
 
 from .models import Alert, Channel, ChannelCount, ChannelLog, SyncEvent
 
 logger = logging.getLogger(__name__)
 
 
-@shared_task(name="sync_channel_fcm_task")
+@shared_task
 def sync_channel_fcm_task(cloud_registration_id, channel_id=None):  # pragma: no cover
     channel = Channel.objects.filter(pk=channel_id).first()
     Channel.sync_channel_fcm(cloud_registration_id, channel)
 
 
-@nonoverlapping_task(name="check_channels_task", lock_key="check_channels")
-def check_channels_task():
+@cron_task()
+def check_channel_alerts():
     """
     Run every 30 minutes.  Checks if any channels who are active have not been seen in that
     time.  Triggers alert in that case
@@ -34,8 +34,8 @@ def check_channels_task():
     Alert.check_alerts()
 
 
-@nonoverlapping_task(name="sync_old_seen_channels_task", lock_key="sync_old_seen_channels")
-def sync_old_seen_channels_task():
+@cron_task()
+def sync_old_seen_channels():
     from temba.channels.types.android import AndroidType
 
     now = timezone.now()
@@ -62,10 +62,10 @@ def interrupt_channel_task(channel_id):
     mailroom.queue_interrupt_channel(channel.org, channel=channel)
 
 
-@nonoverlapping_task(name="trim_sync_events_task")
-def trim_sync_events_task():
+@cron_task()
+def trim_sync_events():
     """
-    Trims old sync events
+    Trims old Android sync events
     """
 
     trim_before = timezone.now() - settings.RETENTION_PERIODS["syncevent"]
@@ -84,8 +84,8 @@ def trim_sync_events_task():
             event.release()
 
 
-@nonoverlapping_task(name="trim_channel_log_task")
-def trim_channel_log_task():
+@cron_task()
+def trim_channel_logs():
     """
     Trims old channel logs
     """
@@ -97,12 +97,12 @@ def trim_channel_log_task():
         ChannelLog.objects.filter(id__in=chunk).delete()
 
 
-@nonoverlapping_task(name="squash_channelcounts", lock_key="squash_channelcounts", lock_timeout=7200)
-def squash_channelcounts():
+@cron_task(lock_timeout=7200)
+def squash_channel_counts():
     ChannelCount.squash()
 
 
-@nonoverlapping_task(name="track_org_channel_counts", lock_key="track_org_channel_counts", lock_timeout=7200)
+@cron_task(lock_timeout=7200)
 def track_org_channel_counts(now=None):
     """
     Run daily, logs to our analytics the number of incoming and outgoing messages/ivr messages per org that had