+++ temba/channels/tests.py
@@ -30,12 +30,12 @@ from temba.utils.models import generate_uuid
 
 from .models import Alert, Channel, ChannelCount, ChannelEvent, ChannelLog, SyncEvent
 from .tasks import (
-    check_channels_task,
-    squash_channelcounts,
-    sync_old_seen_channels_task,
+    check_channel_alerts,
+    squash_channel_counts,
+    sync_old_seen_channels,
     track_org_channel_counts,
-    trim_channel_log_task,
-    trim_sync_events_task,
+    trim_channel_logs,
+    trim_sync_events,
 )
 
 
@@ -1046,7 +1046,7 @@ class ChannelTest(TembaTest):
 
         # make our events old so we can test trimming them
         SyncEvent.objects.all().update(created_on=timezone.now() - timedelta(days=45))
-        trim_sync_events_task()
+        trim_sync_events()
 
         # should be cleared out
         self.assertEqual(1, SyncEvent.objects.all().count())
@@ -1429,7 +1429,7 @@ class ChannelAlertTest(TembaTest):
         self.channel.last_seen = timezone.now() - timedelta(minutes=40)
         self.channel.save()
 
-        check_channels_task()
+        check_channel_alerts()
         self.assertTrue(len(mail.outbox) == 0)
 
         # add alert email, remove org and set last seen to now to force an resolve email to try to send
@@ -1437,7 +1437,7 @@ class ChannelAlertTest(TembaTest):
         self.channel.org = None
         self.channel.last_seen = timezone.now()
         self.channel.save()
-        check_channels_task()
+        check_channel_alerts()
 
         self.assertTrue(len(mail.outbox) == 0)
 
@@ -1448,19 +1448,19 @@ class ChannelSyncTest(TembaTest):
         self.channel.last_seen = timezone.now() - timedelta(days=40)
         self.channel.save()
 
-        sync_old_seen_channels_task()
+        sync_old_seen_channels()
         self.assertFalse(mock_trigger_sync.called)
 
         self.channel.last_seen = timezone.now() - timedelta(minutes=5)
         self.channel.save()
 
-        sync_old_seen_channels_task()
+        sync_old_seen_channels()
         self.assertFalse(mock_trigger_sync.called)
 
         self.channel.last_seen = timezone.now() - timedelta(hours=3)
         self.channel.save()
 
-        sync_old_seen_channels_task()
+        sync_old_seen_channels()
         self.assertTrue(mock_trigger_sync.called)
 
 
@@ -1475,7 +1475,7 @@ class ChannelClaimTest(TembaTest):
         brands = copy.deepcopy(settings.BRANDS)
         brands[0]["from_email"] = "support@mybrand.com"
         with self.settings(BRANDS=brands):
-            check_channels_task()
+            check_channel_alerts()
 
             # should have created one alert
             alert = Alert.objects.get()
@@ -1501,7 +1501,7 @@ class ChannelClaimTest(TembaTest):
             self.assertEqual(mail.outbox[0].from_email, "support@mybrand.com")
 
         # call it again
-        check_channels_task()
+        check_channel_alerts()
 
         # still only one alert
         self.assertEqual(1, Alert.objects.all().count())
@@ -1511,7 +1511,7 @@ class ChannelClaimTest(TembaTest):
         self.channel.last_seen = timezone.now() + timedelta(minutes=5)
         self.channel.save()
 
-        check_channels_task()
+        check_channel_alerts()
 
         # still only one alert, but it is now ended
         alert = Alert.objects.get()
@@ -1553,7 +1553,7 @@ class ChannelClaimTest(TembaTest):
         self.channel.save()
 
         # ok check on our channel
-        check_channels_task()
+        check_channel_alerts()
 
         # we don't have  successfully sent message and we have an alert and only one
         self.assertEqual(Alert.objects.all().count(), 1)
@@ -1577,7 +1577,7 @@ class ChannelClaimTest(TembaTest):
         )
 
         # ok check on our channel
-        check_channels_task()
+        check_channel_alerts()
 
         # if latest_sent_message is after our queued message no alert is created
         self.assertEqual(Alert.objects.all().count(), 1)
@@ -1590,7 +1590,7 @@ class ChannelClaimTest(TembaTest):
         msg1 = self.create_outgoing_msg(contact, "Message One", created_on=two_hours_ago, status="Q")
 
         # check our channel again
-        check_channels_task()
+        check_channel_alerts()
 
         #  no new alert created because we sent one in the past hour
         self.assertEqual(Alert.objects.all().count(), 1)
@@ -1603,7 +1603,7 @@ class ChannelClaimTest(TembaTest):
         alert.save()
 
         # check our channel again
-        check_channels_task()
+        check_channel_alerts()
 
         # this time we have a new alert and should create only one
         self.assertEqual(Alert.objects.all().count(), 2)
@@ -1627,7 +1627,7 @@ class ChannelClaimTest(TembaTest):
 
         # run again, nothing should change
         with self.assertNumQueries(9):
-            check_channels_task()
+            check_channel_alerts()
 
         self.assertEqual(2, Alert.objects.filter(channel=self.channel, ended_on=None).count())
         self.assertTrue(len(mail.outbox) == 2)
@@ -1638,7 +1638,7 @@ class ChannelClaimTest(TembaTest):
         msg1.save(update_fields=("status", "sent_on"))
 
         # run again, our alert should end
-        check_channels_task()
+        check_channel_alerts()
 
         # still only one alert though, and no new email sent, alert must not be ended before one hour
         alert = Alert.objects.all().latest("ended_on")
@@ -1671,7 +1671,7 @@ class ChannelCountTest(TembaTest):
         self.assertDailyCount(self.channel, 2, ChannelCount.INCOMING_MSG_TYPE, msg2.created_on.date())
 
         # squash our counts
-        squash_channelcounts()
+        squash_channel_counts()
 
         # same count
         self.assertDailyCount(self.channel, 2, ChannelCount.INCOMING_MSG_TYPE, msg2.created_on.date())
@@ -1696,7 +1696,7 @@ class ChannelCountTest(TembaTest):
         )
 
         # squash our counts
-        squash_channelcounts()
+        squash_channel_counts()
 
         self.assertDailyCount(self.channel, 1, ChannelCount.OUTGOING_MSG_TYPE, msg3.created_on.date())
         self.assertEqual(ChannelCount.objects.filter(count_type=ChannelCount.SUCCESS_LOG_TYPE).count(), 0)
@@ -2800,7 +2800,7 @@ Content-Type: application/json
             created_on=timezone.now() - timedelta(days=2),
         )
 
-        trim_channel_log_task()
+        trim_channel_logs()
 
         # should only have one log remaining and should be l2
         self.assertEqual(1, ChannelLog.objects.all().count())