+++ temba/flows/tasks.py
@@ -12,7 +12,7 @@ from django.utils.timesince import timesince
 
 from temba.contacts.models import ContactField, ContactGroup
 from temba.utils import chunk_list
-from temba.utils.celery import nonoverlapping_task
+from temba.utils.crons import cron_task
 
 from .models import (
     ExportFlowResultsTask,
@@ -23,6 +23,7 @@ from .models import (
     FlowRevision,
     FlowRun,
     FlowRunCount,
+    FlowRunStatusCount,
     FlowSession,
     FlowStart,
     FlowStartCount,
@@ -32,7 +33,7 @@ FLOW_TIMEOUT_KEY = "flow_timeouts_%y_%m_%d"
 logger = logging.getLogger(__name__)
 
 
-@shared_task(name="update_session_wait_expires")
+@shared_task
 def update_session_wait_expires(flow_id):
     """
     Update the wait_expires_on of any session currently waiting in the given flow
@@ -46,7 +47,7 @@ def update_session_wait_expires(flow_id):
         batch.update(wait_expires_on=F("wait_started_on") + timedelta(minutes=flow.expires_after_minutes))
 
 
-@shared_task(name="export_flow_results_task")
+@shared_task
 def export_flow_results_task(export_id):
     """
     Export a flow to a file and e-mail a link to the user
@@ -57,16 +58,17 @@ def export_flow_results_task(export_id):
     ).get(id=export_id).perform()
 
 
-@nonoverlapping_task(name="squash_flowcounts", lock_timeout=7200)
-def squash_flowcounts():
+@cron_task(lock_timeout=7200)
+def squash_flow_counts():
     FlowNodeCount.squash()
     FlowRunCount.squash()
+    FlowRunStatusCount.squash()
     FlowCategoryCount.squash()
     FlowStartCount.squash()
     FlowPathCount.squash()
 
 
-@nonoverlapping_task(name="trim_flow_revisions")
+@cron_task()
 def trim_flow_revisions():
     start = timezone.now()
 
@@ -85,18 +87,12 @@ def trim_flow_revisions():
     logger.info(f"Trimmed {count} flow revisions since {last_trim} in {elapsed}")
 
 
-@nonoverlapping_task(name="trim_flow_sessions_and_starts")
-def trim_flow_sessions_and_starts():
-    num_sessions = _trim_flow_sessions()
-    num_starts = _trim_flow_starts()
-
-    return {"sessions": num_sessions, "starts": num_starts}
-
-
-def _trim_flow_sessions() -> int:
+@cron_task()
+def trim_flow_sessions():
     """
     Cleanup old flow sessions
     """
+
     trim_before = timezone.now() - settings.RETENTION_PERIODS["flowsession"]
     num_deleted = 0
 
@@ -111,13 +107,15 @@ def _trim_flow_sessions() -> int:
         FlowSession.objects.filter(id__in=session_ids).delete()
         num_deleted += len(session_ids)
 
-    return num_deleted
+    return {"deleted": num_deleted}
 
 
-def _trim_flow_starts() -> int:
+@cron_task()
+def trim_flow_starts() -> int:
     """
     Cleanup completed non-user created flow starts
     """
+
     trim_before = timezone.now() - settings.RETENTION_PERIODS["flowstart"]
     num_deleted = 0
 
@@ -147,4 +145,4 @@ def _trim_flow_starts() -> int:
         FlowStart.objects.filter(id__in=start_ids).delete()
         num_deleted += len(start_ids)
 
-    return num_deleted
+    return {"deleted": num_deleted}