+++ temba/flows/tests.py
@@ -27,7 +27,7 @@ from temba.globals.models import Global
 from temba.mailroom import FlowValidationException
 from temba.orgs.integrations.dtone import DTOneType
 from temba.templates.models import Template, TemplateTranslation
-from temba.tests import AnonymousOrg, CRUDLTestMixin, MigrationTest, MockResponse, TembaTest, matchers, mock_mailroom
+from temba.tests import AnonymousOrg, CRUDLTestMixin, MockResponse, TembaTest, matchers, mock_mailroom
 from temba.tests.engine import MockSessionWriter
 from temba.tests.s3 import MockS3Client, jsonlgz_encode
 from temba.tickets.models import Ticketer
@@ -45,7 +45,7 @@ from .models import (
     FlowPathCount,
     FlowRevision,
     FlowRun,
-    FlowRunCount,
+    FlowRunStatusCount,
     FlowSession,
     FlowStart,
     FlowStartCount,
@@ -53,7 +53,13 @@ from .models import (
     FlowVersionConflictException,
     get_flow_user,
 )
-from .tasks import squash_flowcounts, trim_flow_revisions, trim_flow_sessions_and_starts, update_session_wait_expires
+from .tasks import (
+    squash_flow_counts,
+    trim_flow_revisions,
+    trim_flow_sessions,
+    trim_flow_starts,
+    update_session_wait_expires,
+)
 from .views import FlowCRUDL
 
 
@@ -431,7 +437,11 @@ class FlowTest(TembaTest):
             visited,
         )
         self.assertEqual(
-            {"total": 1, "active": 1, "completed": 0, "expired": 0, "interrupted": 0, "failed": 0, "completion": 0},
+            {
+                "total": 1,
+                "status": {"active": 0, "waiting": 1, "completed": 0, "expired": 0, "interrupted": 0, "failed": 0},
+                "completion": 0,
+            },
             flow.get_run_stats(),
         )
 
@@ -516,7 +526,11 @@ class FlowTest(TembaTest):
             visited,
         )
         self.assertEqual(
-            {"total": 2, "active": 2, "completed": 0, "expired": 0, "interrupted": 0, "failed": 0, "completion": 0},
+            {
+                "total": 2,
+                "status": {"active": 0, "waiting": 2, "completed": 0, "expired": 0, "interrupted": 0, "failed": 0},
+                "completion": 0,
+            },
             flow.get_run_stats(),
         )
 
@@ -567,12 +581,16 @@ class FlowTest(TembaTest):
 
         # half of our flows are now complete
         self.assertEqual(
-            {"total": 2, "active": 1, "completed": 1, "expired": 0, "interrupted": 0, "failed": 0, "completion": 50},
+            {
+                "total": 2,
+                "status": {"active": 0, "waiting": 1, "completed": 1, "expired": 0, "interrupted": 0, "failed": 0},
+                "completion": 50,
+            },
             flow.get_run_stats(),
         )
 
         # check squashing doesn't change anything
-        squash_flowcounts()
+        squash_flow_counts()
 
         (active, visited) = flow.get_activity()
 
@@ -591,7 +609,11 @@ class FlowTest(TembaTest):
             visited,
         )
         self.assertEqual(
-            {"total": 2, "active": 1, "completed": 1, "expired": 0, "interrupted": 0, "failed": 0, "completion": 50},
+            {
+                "total": 2,
+                "status": {"active": 0, "waiting": 1, "completed": 1, "expired": 0, "interrupted": 0, "failed": 0},
+                "completion": 50,
+            },
             flow.get_run_stats(),
         )
         self.assertEqual(
@@ -632,7 +654,11 @@ class FlowTest(TembaTest):
             visited,
         )
         self.assertEqual(
-            {"total": 1, "active": 1, "completed": 0, "expired": 0, "interrupted": 0, "failed": 0, "completion": 0},
+            {
+                "total": 1,
+                "status": {"active": 0, "waiting": 1, "completed": 0, "expired": 0, "interrupted": 0, "failed": 0},
+                "completion": 0,
+            },
             flow.get_run_stats(),
         )
         self.assertEqual(
@@ -682,7 +708,11 @@ class FlowTest(TembaTest):
             visited,
         )
         self.assertEqual(
-            {"total": 1, "active": 0, "completed": 1, "expired": 0, "interrupted": 0, "failed": 0, "completion": 100},
+            {
+                "total": 1,
+                "status": {"active": 0, "waiting": 0, "completed": 1, "expired": 0, "interrupted": 0, "failed": 0},
+                "completion": 100,
+            },
             flow.get_run_stats(),
         )
 
@@ -706,7 +736,11 @@ class FlowTest(TembaTest):
             visited,
         )
         self.assertEqual(
-            {"total": 0, "active": 0, "completed": 0, "expired": 0, "interrupted": 0, "failed": 0, "completion": 0},
+            {
+                "total": 0,
+                "status": {"active": 0, "waiting": 0, "completed": 0, "expired": 0, "interrupted": 0, "failed": 0},
+                "completion": 0,
+            },
             flow.get_run_stats(),
         )
         self.assertEqual(
@@ -762,7 +796,11 @@ class FlowTest(TembaTest):
             visited,
         )
         self.assertEqual(
-            {"total": 1, "active": 1, "completed": 0, "expired": 0, "interrupted": 0, "failed": 0, "completion": 0},
+            {
+                "total": 1,
+                "status": {"active": 0, "waiting": 1, "completed": 0, "expired": 0, "interrupted": 0, "failed": 0},
+                "completion": 0,
+            },
             flow.get_run_stats(),
         )
 
@@ -789,7 +827,11 @@ class FlowTest(TembaTest):
             visited,
         )
         self.assertEqual(
-            {"total": 1, "active": 0, "completed": 0, "expired": 1, "interrupted": 0, "failed": 0, "completion": 0},
+            {
+                "total": 1,
+                "status": {"active": 0, "waiting": 0, "completed": 0, "expired": 1, "interrupted": 0, "failed": 0},
+                "completion": 0,
+            },
             flow.get_run_stats(),
         )
 
@@ -813,7 +855,11 @@ class FlowTest(TembaTest):
 
         self.assertEqual({color_split["uuid"]: 1}, active)
         self.assertEqual(
-            {"total": 2, "active": 1, "completed": 0, "expired": 1, "interrupted": 0, "failed": 0, "completion": 0},
+            {
+                "total": 2,
+                "status": {"active": 0, "waiting": 1, "completed": 0, "expired": 1, "interrupted": 0, "failed": 0},
+                "completion": 0,
+            },
             flow.get_run_stats(),
         )
 
@@ -826,30 +872,13 @@ class FlowTest(TembaTest):
 
         self.assertEqual({}, active)
         self.assertEqual(
-            {"total": 2, "active": 0, "completed": 0, "expired": 1, "interrupted": 1, "failed": 0, "completion": 0},
+            {
+                "total": 2,
+                "status": {"active": 0, "waiting": 0, "completed": 0, "expired": 1, "interrupted": 1, "failed": 0},
+                "completion": 0,
+            },
             flow.get_run_stats(),
         )
-
-    def test_squash_counts(self):
-        flow = self.get_flow("favorites")
-        flow2 = self.get_flow("pick_a_number")
-
-        FlowRunCount.objects.create(flow=flow, count=2, exit_type=None)
-        FlowRunCount.objects.create(flow=flow, count=1, exit_type=None)
-        FlowRunCount.objects.create(flow=flow, count=3, exit_type="E")
-        FlowRunCount.objects.create(flow=flow2, count=10, exit_type="I")
-        FlowRunCount.objects.create(flow=flow2, count=-1, exit_type="I")
-
-        squash_flowcounts()
-        self.assertEqual(FlowRunCount.objects.all().count(), 3)
-        self.assertEqual(FlowRunCount.get_totals(flow2), {"I": 9})
-        self.assertEqual(FlowRunCount.get_totals(flow), {None: 3, "E": 3})
-
-        max_id = FlowRunCount.objects.all().order_by("-id").first().id
-
-        # no-op this time
-        squash_flowcounts()
-        self.assertEqual(max_id, FlowRunCount.objects.all().order_by("-id").first().id)
 
     def test_category_counts(self):
         def assertCount(counts, result_key, category_name, truth):
@@ -1073,7 +1102,7 @@ class FlowTest(TembaTest):
         self.assertEqual(2, FlowCategoryCount.objects.filter(category_name="Blue", result_name="Color").count())
         FlowCategoryCount.objects.get(category_name="Blue", result_name="Color", result_key="color", count=-1)
 
-    def test_flow_start_counts(self):
+    def test_start_counts(self):
         # create start for 10 contacts
         flow = self.create_flow("Test")
         start = FlowStart.objects.create(org=self.org, flow=flow, created_by=self.admin)
@@ -3139,9 +3168,10 @@ class FlowCRUDLTest(TembaTest, CRUDLTestMixin):
             # and some charts
             response = self.client.get(reverse("flows.flow_activity_chart", args=[flow.id]))
 
-            # we have two active runs, one failed run
+            # we have two waiting runs, one failed run
             self.assertEqual(response.context["failed"], 1)
-            self.assertEqual(response.context["active"], 2)
+            self.assertEqual(response.context["active"], 0)
+            self.assertEqual(response.context["waiting"], 2)
             self.assertEqual(response.context["completed"], 0)
             self.assertEqual(response.context["expired"], 0)
             self.assertEqual(response.context["interrupted"], 0)
@@ -3176,11 +3206,12 @@ class FlowCRUDLTest(TembaTest, CRUDLTestMixin):
             self.assertEqual(1, len(response.context["runs"]))
             self.assertContains(response, "Jimmy")
 
-            # now only one active, one completed, one failed and 5 total responses
+            # now only one waiting, one completed, one failed and 5 total responses
             response = self.client.get(reverse("flows.flow_activity_chart", args=[flow.id]))
 
             self.assertEqual(response.context["failed"], 1)
-            self.assertEqual(response.context["active"], 1)
+            self.assertEqual(response.context["active"], 0)
+            self.assertEqual(response.context["waiting"], 1)
             self.assertEqual(response.context["completed"], 1)
             self.assertEqual(response.context["expired"], 0)
             self.assertEqual(response.context["interrupted"], 0)
@@ -3554,6 +3585,104 @@ class FlowRunTest(TembaTest):
 
         self.contact = self.create_contact("Ben Haggerty", phone="+250788123123")
 
+    def test_status_counts(self):
+        contact = self.create_contact("Bob", phone="+1234567890")
+        session = FlowSession.objects.create(
+            uuid=uuid4(),
+            org=self.org,
+            contact=self.contact,
+            status=FlowSession.STATUS_WAITING,
+            output_url="http://sessions.com/123.json",
+            created_on=timezone.now(),
+            wait_started_on=timezone.now(),
+            wait_expires_on=timezone.now() + timedelta(days=7),
+            wait_resume_on_expire=False,
+        )
+
+        def create_runs(flow_status_pairs: tuple) -> list:
+            runs = []
+            for flow, status in flow_status_pairs:
+                runs.append(
+                    FlowRun(
+                        uuid=uuid4(),
+                        org=self.org,
+                        session=session,
+                        flow=flow,
+                        contact=contact,
+                        status=status,
+                        created_on=timezone.now(),
+                        modified_on=timezone.now(),
+                        exited_on=timezone.now() if status not in ("A", "W") else None,
+                    )
+                )
+            return FlowRun.objects.bulk_create(runs)
+
+        flow1 = self.create_flow("Test 1")
+        flow2 = self.create_flow("Test 2")
+
+        runs1 = create_runs(
+            (
+                (flow1, FlowRun.STATUS_ACTIVE),
+                (flow2, FlowRun.STATUS_WAITING),
+                (flow1, FlowRun.STATUS_ACTIVE),
+                (flow2, FlowRun.STATUS_WAITING),
+                (flow1, FlowRun.STATUS_WAITING),
+                (flow1, FlowRun.STATUS_COMPLETED),
+            )
+        )
+
+        self.assertEqual(
+            {(flow1, "A"): 2, (flow2, "W"): 2, (flow1, "W"): 1, (flow1, "C"): 1},
+            {(c.flow, c.status): c.count for c in FlowRunStatusCount.objects.all()},
+        )
+        self.assertEqual({"A": 2, "W": 1, "C": 1}, FlowRunStatusCount.get_totals(flow1))
+        self.assertEqual({"W": 2}, FlowRunStatusCount.get_totals(flow2))
+
+        # no difference after squashing
+        squash_flow_counts()
+
+        self.assertEqual({"A": 2, "W": 1, "C": 1}, FlowRunStatusCount.get_totals(flow1))
+        self.assertEqual({"W": 2}, FlowRunStatusCount.get_totals(flow2))
+
+        runs2 = create_runs(
+            (
+                (flow1, FlowRun.STATUS_ACTIVE),
+                (flow1, FlowRun.STATUS_ACTIVE),
+                (flow2, FlowRun.STATUS_EXPIRED),
+            )
+        )
+
+        self.assertEqual({"A": 4, "W": 1, "C": 1}, FlowRunStatusCount.get_totals(flow1))
+        self.assertEqual({"W": 2, "X": 1}, FlowRunStatusCount.get_totals(flow2))
+
+        # bulk update runs like they're being interrupted
+        FlowRun.objects.filter(id__in=[r.id for r in runs1]).update(
+            status=FlowRun.STATUS_INTERRUPTED, exited_on=timezone.now()
+        )
+
+        self.assertEqual({"A": 2, "W": 0, "C": 0, "I": 4}, FlowRunStatusCount.get_totals(flow1))
+        self.assertEqual({"W": 0, "X": 1, "I": 2}, FlowRunStatusCount.get_totals(flow2))
+
+        # no difference after squashing
+        squash_flow_counts()
+
+        self.assertEqual({"A": 2, "W": 0, "C": 0, "I": 4}, FlowRunStatusCount.get_totals(flow1))
+        self.assertEqual({"W": 0, "X": 1, "I": 2}, FlowRunStatusCount.get_totals(flow2))
+
+        # do manual deletion of some runs
+        FlowRun.objects.filter(id__in=[r.id for r in runs2]).update(delete_from_results=True)
+        FlowRun.objects.filter(id__in=[r.id for r in runs2]).delete()
+
+        self.assertEqual({"A": 0, "W": 0, "C": 0, "I": 4}, FlowRunStatusCount.get_totals(flow1))
+        self.assertEqual({"W": 0, "X": 0, "I": 2}, FlowRunStatusCount.get_totals(flow2))
+
+        # do archival deletion of the rest
+        FlowRun.objects.filter(id__in=[r.id for r in runs1]).delete()
+
+        # status counts are unchanged
+        self.assertEqual({"A": 0, "W": 0, "C": 0, "I": 4}, FlowRunStatusCount.get_totals(flow1))
+        self.assertEqual({"W": 0, "X": 0, "I": 2}, FlowRunStatusCount.get_totals(flow2))
+
     def test_as_archive_json(self):
         flow = self.get_flow("color_v13")
         flow_nodes = flow.get_definition()["nodes"]
@@ -3724,11 +3853,14 @@ class FlowRunTest(TembaTest):
                 "start_count": 1,  # unchanged
                 "run_count": {
                     "total": 0,
-                    "active": 0,
-                    "completed": 0,
-                    "expired": 0,
-                    "interrupted": 0,
-                    "failed": 0,
+                    "status": {
+                        "active": 0,
+                        "waiting": 0,
+                        "completed": 0,
+                        "expired": 0,
+                        "interrupted": 0,
+                        "failed": 0,
+                    },
                     "completion": 0,
                 },
             },
@@ -3745,11 +3877,14 @@ class FlowRunTest(TembaTest):
                 "start_count": 1,  # unchanged
                 "run_count": {
                     "total": 0,
-                    "active": 0,
-                    "completed": 0,
-                    "expired": 0,
-                    "interrupted": 0,
-                    "failed": 0,
+                    "status": {
+                        "active": 0,
+                        "waiting": 0,
+                        "completed": 0,
+                        "expired": 0,
+                        "interrupted": 0,
+                        "failed": 0,
+                    },
                     "completion": 0,
                 },
             },
@@ -3767,11 +3902,14 @@ class FlowRunTest(TembaTest):
                 "start_count": 1,  # unchanged
                 "run_count": {  # unchanged
                     "total": 1,
-                    "active": 0,
-                    "completed": 1,
-                    "expired": 0,
-                    "interrupted": 0,
-                    "failed": 0,
+                    "status": {
+                        "active": 0,
+                        "waiting": 0,
+                        "completed": 1,
+                        "expired": 0,
+                        "interrupted": 0,
+                        "failed": 0,
+                    },
                     "completion": 100,
                 },
             },
@@ -3906,7 +4044,7 @@ class FlowSessionTest(TembaTest):
         run2.session.ended_on = timezone.now()
         run2.session.save(update_fields=("status", "ended_on"))
 
-        trim_flow_sessions_and_starts()
+        trim_flow_sessions()
 
         run1, run2, run3, run4 = FlowRun.objects.order_by("id")
 
@@ -3968,7 +4106,7 @@ class FlowStartTest(TembaTest):
         create_start(None, FlowStart.TYPE_FLOW_ACTION, FlowStart.STATUS_COMPLETE, date2, contacts=[contact])
         create_start(None, FlowStart.TYPE_TRIGGER, FlowStart.STATUS_FAILED, date2, groups=[group])
 
-        trim_flow_sessions_and_starts()
+        trim_flow_starts()
 
         # check that related objects still exist!
         contact.refresh_from_db()
@@ -5694,39 +5832,3 @@ class FlowRevisionTest(TembaTest):
         trim_flow_revisions()
         self.assertEqual(2, FlowRevision.objects.filter(flow=clinic).count())
         self.assertEqual(31, FlowRevision.objects.filter(flow=color).count())
-
-
-class FlattenFlowLabelsTest(MigrationTest):
-    app = "flows"
-    migrate_from = "0301_exportflowresultstask_with_groups"
-    migrate_to = "0302_flatten_flow_labels"
-
-    def setUpBeforeMigration(self, apps):
-        self.flow1 = self.create_flow("Flow 1")
-        self.flow2 = self.create_flow("Flow 2")
-        self.flow3 = self.create_flow("Flow 3")
-        self.flow4 = self.create_flow("Flow 4")
-
-        def create_label(name, parent, flows):
-            label = FlowLabel.objects.create(
-                org=self.org, name=name, parent=parent, created_by=self.admin, modified_by=self.admin
-            )
-            label.flows.add(*flows)
-            return label
-
-        self.label1 = create_label("Foo", parent=None, flows=[self.flow1])
-        self.label1_child1 = create_label("1", parent=self.label1, flows=[self.flow1, self.flow2])
-        self.label1_child2 = create_label("2", parent=self.label1, flows=[self.flow3])
-        self.label2 = create_label("Bar", parent=None, flows=[self.flow4])
-
-    def test_migration(self):
-        def assert_label(lbl, expected_name, expected_flows):
-            lbl.refresh_from_db()
-            self.assertEqual(expected_name, lbl.name)
-            self.assertEqual(expected_flows, set(lbl.flows.all()))
-            self.assertIsNone(lbl.parent)
-
-        assert_label(self.label1, "Foo", {self.flow1, self.flow2, self.flow3})
-        assert_label(self.label1_child1, "Foo > 1", {self.flow1, self.flow2})
-        assert_label(self.label1_child2, "Foo > 2", {self.flow3})
-        assert_label(self.label2, "Bar", {self.flow4})