+++ temba/orgs/tasks.py
@@ -11,18 +11,18 @@ from temba.flows.models import ExportFlowResultsTask
 from temba.flows.tasks import export_flow_results_task
 from temba.msgs.models import ExportMessagesTask
 from temba.msgs.tasks import export_messages_task
-from temba.utils.celery import nonoverlapping_task
+from temba.utils.crons import cron_task
 
 from .models import Invitation, Org
 
 
-@shared_task(name="send_invitation_email_task")
+@shared_task
 def send_invitation_email_task(invitation_id):
     invitation = Invitation.objects.get(pk=invitation_id)
     invitation.send_email()
 
 
-@shared_task(name="normalize_contact_tels_task")
+@shared_task
 def normalize_contact_tels_task(org_id):
     org = Org.objects.get(id=org_id)
 
@@ -33,7 +33,7 @@ def normalize_contact_tels_task(org_id):
             urn.ensure_number_normalization(org.default_country_code)
 
 
-@nonoverlapping_task(name="resume_failed_tasks", lock_key="resume_failed_tasks", lock_timeout=7200)
+@cron_task(lock_timeout=7200)
 def resume_failed_tasks():
     now = timezone.now()
     window = now - timedelta(hours=1)
@@ -57,8 +57,8 @@ def resume_failed_tasks():
         export_messages_task.delay(msg_export.pk)
 
 
-@nonoverlapping_task(name="delete_orgs_task", lock_key="delete_orgs_task", lock_timeout=7200)
-def delete_orgs_task():
+@cron_task(lock_timeout=7200)
+def delete_released_orgs():
     # for each org that was released over 7 days ago, delete it for real
     week_ago = timezone.now() - timedelta(days=Org.DELETE_DELAY_DAYS)
     for org in Org.objects.filter(is_active=False, released_on__lt=week_ago, deleted_on=None):