+++ temba/utils/tests.py
@@ -8,6 +8,7 @@ from unittest.mock import patch
 
 import pytz
 from celery.app.task import Task
+from django_redis import get_redis_connection
 
 from django.conf import settings
 from django.forms import ValidationError
@@ -23,7 +24,7 @@ from temba.utils import json, uuid
 from temba.utils.templatetags.temba import format_datetime, icon
 
 from . import chunk_list, countries, format_number, languages, percentage, redact, sizeof_fmt, str_to_bool
-from .celery import nonoverlapping_task
+from .crons import clear_cron_stats, cron_task
 from .dates import date_range, datetime_to_str, datetime_to_timestamp, timestamp_to_datetime
 from .email import is_valid_address, send_simple_email
 from .fields import NameValidator, validate_external_url
@@ -536,22 +537,26 @@ class JsonTest(TembaTest):
             json.dumps(dict(foo=Exception("invalid")))
 
 
-class CeleryTest(TembaTest):
+class CronsTest(TembaTest):
     @patch("redis.client.StrictRedis.lock")
     @patch("redis.client.StrictRedis.get")
-    def test_nonoverlapping_task(self, mock_redis_get, mock_redis_lock):
+    def test_cron_task(self, mock_redis_get, mock_redis_lock):
+        clear_cron_stats()
+
         mock_redis_get.return_value = None
         task_calls = []
 
-        @nonoverlapping_task()
+        @cron_task()
         def test_task1(foo, bar):
             task_calls.append("1-%d-%d" % (foo, bar))
+            return {"foo": 1}
 
-        @nonoverlapping_task(name="task2", time_limit=100)
+        @cron_task(name="task2", time_limit=100)
         def test_task2(foo, bar):
             task_calls.append("2-%d-%d" % (foo, bar))
+            return 1234
 
-        @nonoverlapping_task(name="task3", time_limit=100, lock_key="test_key", lock_timeout=55)
+        @cron_task(name="task3", time_limit=100, lock_timeout=55)
         def test_task3(foo, bar):
             task_calls.append("3-%d-%d" % (foo, bar))
 
@@ -569,12 +574,21 @@ class CeleryTest(TembaTest):
 
         mock_redis_get.assert_any_call("celery-task-lock:test_task1")
         mock_redis_get.assert_any_call("celery-task-lock:task2")
-        mock_redis_get.assert_any_call("test_key")
+        mock_redis_get.assert_any_call("celery-task-lock:task3")
         mock_redis_lock.assert_any_call("celery-task-lock:test_task1", timeout=900)
         mock_redis_lock.assert_any_call("celery-task-lock:task2", timeout=100)
-        mock_redis_lock.assert_any_call("test_key", timeout=55)
+        mock_redis_lock.assert_any_call("celery-task-lock:task3", timeout=55)
 
         self.assertEqual(task_calls, ["1-11-12", "2-21-22", "3-31-32"])
+
+        r = get_redis_connection()
+        self.assertEqual({b"test_task1", b"task2", b"task3"}, set(r.hkeys("cron_stats:last_start")))
+        self.assertEqual({b"test_task1", b"task2", b"task3"}, set(r.hkeys("cron_stats:last_time")))
+        self.assertEqual(
+            {b"test_task1": b'{"foo": 1}', b"task2": b"1234", b"task3": b"null"}, r.hgetall("cron_stats:last_result")
+        )
+        self.assertEqual({b"test_task1": b"1", b"task2": b"1", b"task3": b"1"}, r.hgetall("cron_stats:call_count"))
+        self.assertEqual({b"test_task1", b"task2", b"task3"}, set(r.hkeys("cron_stats:total_time")))
 
         # simulate task being already running
         mock_redis_get.reset_mock()