+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+v7.5.131 (2022-12-09)
+-------------------------
+ * Replace .gauge on analytics backend with .gauges which allows backends to send guage values in bulk
+ * Remove celery auto discovery for jiochat and wechat tasks which were removed
+
+v7.5.130 (2022-12-09)
+-------------------------
+ * Record cron time in analytics
+
+v7.5.129 (2022-12-08)
+-------------------------
+ * Cleanup cron task names
+ * Split task to trim starts and sessions into two separate tasks
+ * Expose all status counts on flows endpoint
+ * Read from FlowRunStatusCount instead of FlowRunCount
+ * Track flow start counts in statement rather than row level trigger
+
+v7.5.128 (2022-12-07)
+-------------------------
+ * Record cron task last stats in redis
+ * Switch from flake8 to ruff
+ * Add data migration to convert exit_type counts to status counts
+
+v7.5.127 (2022-12-07)
+-------------------------
+ * Fix counts for triggers on the menu
+
+v7.5.126 (2022-12-06)
+-------------------------
+ * Add new count model for run statuses managed by by-statement db triggers
+
+v7.5.125 (2022-12-05)
+-------------------------
+ * Tweak index used to find messages to retry so that it includes PENDING messages
+
 v7.5.124 (2022-12-05)
 -------------------------
  * Update to latest components